# api-sys_produtos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Produtos-api Documentation",
        "description": "api-sys_produtos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8989",
            "description": "local server"
        }
    ],
    "paths": {
        "/api/categoria/get/{id}": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Get Categoria",
                "description": "Categorias Controller",
                "operationId": "ba5e87a6c2c86963d243acf50a5f2414",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/categoria/create": {
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Create Categoria",
                "description": "Categorias Controller",
                "operationId": "5e2c2341a09e3bf71aa43435c70a3840",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/categoria/update": {
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Update Categoria",
                "description": "Categorias Controller",
                "operationId": "269b79bb094630057d5b64bf615f4c3b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth login",
                "description": "Auth Controller",
                "operationId": "2ede96b91e694d9d6323469efacc087e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth logout",
                "description": "Auth Controller",
                "operationId": "f1edef7f79a0461455ab182363b1fd37",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}